Refactoring of scraping gcf
